expand active item/category
There isn't really a good way to deal with situations where there is more than one category for a post. For now, the script just expands/highlights the first occurrence. This will probably work most of the time, since the only multiple instance posts seem to be OS related. Showing the non-OS category is probably better. Ideally, we should come up with a more concise method of organization where there are no duplicates.